fixing the issue of where to add the login section
because im using floats, im finding that if i want to add anything above the side bar, i was just moving both the side bar and the sections to the right.

however if i add the login section to the sidebar div already created i get the effect im looking for without having to leave a gap under the image . i have also fiex the sidebar inplace and removed the scroll feature.